+++ eric_sse/queue.py
@@ -1,14 +1,14 @@
 from abc import ABC, abstractmethod
 from threading import Lock
 
-from eric_sse.message import Message
+from eric_sse.message import MessageContract
 from eric_sse.exception import NoMessagesException
 
 
 class Queue(ABC):
     """Abstract base class for queues (FIFO)"""
     @abstractmethod
-    def pop(self) -> Message:
+    def pop(self) -> MessageContract:
         """
         Next message from the queue.
 
@@ -17,7 +17,7 @@ class Queue(ABC):
         ...
 
     @abstractmethod
-    def push(self, message: Message) -> None:
+    def push(self, message: MessageContract) -> None:
         ...
 
     @abstractmethod
@@ -27,9 +27,9 @@ class Queue(ABC):
 
 class InMemoryQueue(Queue):
     def __init__(self):
-        self.__messages: list[Message] = []
+        self.__messages: list[MessageContract] = []
 
-    def pop(self) -> Message:
+    def pop(self) -> MessageContract:
         """
         """
         try:
@@ -38,7 +38,7 @@ class InMemoryQueue(Queue):
         except IndexError:
             raise NoMessagesException
 
-    def push(self, message: Message) -> None:
+    def push(self, message: MessageContract) -> None:
         self.__messages.append(message)
 
     def delete(self) -> None: