+++ docs_markdown/changelog.md
@@ -2,6 +2,10 @@
 
 # Changelog
 
+0.6.4.1
+
+* Changed method’s signature from Message to MessageContract
+
 0.6.4
 
 Breaking: